code push

+++ src/container/pages/meeting/scedulemeeting/Actions/Actions.js
@@ -117,15 +117,6 @@ const Actions = ({
     dispatch(getTodoStatus(navigate, t));
   }, []);
 
-  useEffect(() => {
-    if (
-      toDoListReducer.SocketTodoActivityData !== null &&
-      toDoListReducer.SocketTodoActivityData !== undefined
-    ) {
-      setActionsRows([toDoListReducer.SocketTodoActivityData, ...actionsRows]);
-    }
-  }, [toDoListReducer.SocketTodoActivityData]);
-
   // for pagination in Create Task
   const handleForPagination = (current, pageSize) => {
     let data = {
@@ -167,6 +158,22 @@ const Actions = ({
     }
     setStatusOptions(optionsArr);
   }, [todoStatus]);
+
+  // Remove task from mqtt response
+  useEffect(() => {
+    try {
+      if (toDoListReducer.socketTodoStatusData !== null) {
+        let payloadData = toDoListReducer.socketTodoStatusData;
+        if (payloadData.todoStatusID === 6) {
+          setActionsRows((rowsData) => {
+            return rowsData.filter((newData, index) => {
+              return newData.pK_TID !== payloadData.todoid;
+            });
+          });
+        }
+      }
+    } catch {}
+  }, [toDoListReducer.socketTodoStatusData]);
 
   useEffect(() => {
     if (removeTodo !== 0) {
@@ -422,16 +429,18 @@ const Actions = ({
       key: "RedCrossIcon",
       width: "50px",
       render: (text, record) => {
-        return (
-          <i>
-            <img
-              alt={"Cross"}
-              src={del}
-              className={styles["action-delete-cursor"]}
-              onClick={() => deleteActionHandler(record)}
-            />
-          </i>
-        );
+        if (Number(record?.taskCreator?.pK_UID) === Number(userID)) {
+          return (
+            <i>
+              <img
+                alt={"Cross"}
+                src={del}
+                className={styles["action-delete-cursor"]}
+                onClick={() => deleteActionHandler(record)}
+              />
+            </i>
+          );
+        }
       },
     },
   ];
